Add ljdump2jrnl
This script takes the output of ljdump (which are XML files per
entry) and makes a set of files that can be imported by running
jrnl < file to automatically set the date, title, and merge the
comments in the entry in a pleasing way.

Apologies for the hardcoded paths.  I'm feeling too lazy.